+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.15] - 2020-06-22
+### Changed
+- Keyed each blocks
+### Added
+- Thing.svelte
+
 ## [0.1.14] - 2020-06-11
 ### Changed
 - Loop over lists of data